ASSISTIVE_VISION_PRO/requirements.txt
ultralytics==8.3.0
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
pyttsx3==2.90
torch==2.4.1
torchvision==0.19.1
